[Tempest][Tobiko] clouds.yaml mounted on the proper path
When either a tempest or a tobiko container is run, the clouds.yaml file
could be mounted directly on its proper path
`$HOME/.config/openstack/clouds.yaml` (new behavior implemented in
ci-framework) or on the `external_files` directory (legacy behavior).

+++ container-images/tcib/base/os/tempest/run_tempest.sh
@@ -239,7 +239,7 @@ function generate_test_results {
 
 export OS_CLOUD=default
 
-if [ ! -z ${USE_EXTERNAL_FILES} ]; then
+if [ ! -z ${USE_EXTERNAL_FILES} ] && [ ! -f $HOME/.config/openstack/clouds.yaml ]; then
     mkdir -p $HOME/.config/openstack
     cp ${TEMPEST_PATH}clouds.yaml $HOME/.config/openstack/clouds.yaml
 fi
